+++ src/redux/transactionsTable/transactions-selectors.js
