+++ src/views/MainView.js
@@ -1,10 +1,10 @@
 import { useDispatch, useSelector } from 'react-redux';
 import { useEffect } from 'react';
-import Table from '../components/Table';
-import Balance from '../components/Table/Balance';
-import { fetchTransactions } from '../redux/table/table-operations';
+import Table from '../components/TransactionsTable/Table';
+import Balance from '../components/Balance/Balance';
+import { fetchTransactions } from '../redux/transactionsTable/transactions-operations';
 
-export default function TableView() {
+export default function MainView() {
   const dispatch = useDispatch();
   useEffect(() => {
     dispatch(fetchTransactions());
